+++ docs/site/.gitignore
@@ -2,6 +2,9 @@ content/openapi/artifacts/*
 !content/openapi/artifacts/index.mdx
 !content/openapi/index.mdx
 
+# Generated data files
+content/examples-data.json
+
 .source
 /public/*.st
 /public/*.toml
